requirements.txt
fitz
PyMuPDF
numpy
Pillow
PyPDF2
streamlit
pdf2image
poppler

